refactor load test to share jobs
